refactor: split control_plane.py under the 500-line cap
Drain raes_runtime/control_plane.py from 707 to 338 lines by extracting two
cohesive clusters into sibling modules, matching the existing control_plane_*
module and mixin layout:

- control_plane_submission.py: plan submission-admission diagnostics.
- control_plane_workflow_control.py: WorkflowControlMixin with workflow
  cancellation and timeout reconciliation.

Reuse the canonical _utc_now from control_plane_execution instead of the local
duplicate, and drop the file from tools/policy/oversized_allowlist.yaml now
that it clears the ADR-015 cap.

Pure refactor with no behavior change: RuntimeControlPlane and every external
import site resolve identically, no public API changed, and the full nox verify
suite (hygiene, policy, lint, contracts, pytest, integration, docs) is green.
